refactor(architecture): migrating acopled modules to its own repository
